refactor: move shim command to internal/cli (#529 slice 7, final)
Lift-and-shift the `shim` command (the in-process MySQL-protocol
time-travel server for the _flashback/_snapshot/_diff virtual schemas)
and its two test files from cmd/bintrail (package main) into internal/cli
(package cli), registering it via AddReadCommands so the bintrail and the
upcoming bintrail-pg binaries share one definition.

This is the last read-plane command to migrate, completing #529. shim is a
long-running server rather than a one-shot query, but it serves the same
shared index, so it belongs on the read/serve plane alongside
query/recover/reconstruct/status. It links go-mysql/server for its wire
protocol — that's fine in internal/cli (the read-layer Event guard and the
console uifree guard both exclude shim by design, #528).

init-shim (the ProxySQL/MySQL shim.yaml generator) is intentionally NOT
moved: it's source-specific config tooling, not a read-plane command.

Added TestShimCmd_registered (sibling of the other *_registered tests) and
grew the read_commands_test.go binary-wiring guard to include shim. Behavior
unchanged: same flags, same server, same output.

Closes #529.
