+++ cmd/bintrail/read_commands_test.go
@@ -12,7 +12,7 @@ import "testing"
 //
 // Grow `want` as more read commands migrate into internal/cli (#529).
 func TestReadCommandsWiredOnRoot(t *testing.T) {
-	want := []string{"status", "query", "recover", "reconstruct"}
+	want := []string{"status", "query", "recover", "reconstruct", "shim"}
 
 	have := make(map[string]bool)
 	for _, c := range rootCmd.Commands() {
